fixes #8 round function to return correct data type

+++ src/main/java/Team4450/Lib/Util.java
@@ -43,20 +43,20 @@ public class Util
 
 	/**
 	 * Logging class for use by other classes to log though this custom logging scheme. All
-	 * logging should be done by calls to methods on this class instance or with the 
+	 * logging should be done by calls to methods on this class instance or with the
 	 * convenience methods elsewhere in the Util class.
 	 */
 	public final static Logger logger = Logger.getGlobal();
-	
+
 	private static double timeMarker = 0;
-	
+
 	// Private constructor means this class cannot be instantiated. All access is static.
-	
+
 	private Util()
 	{
-		
+
 	}
-	
+
 	/**
 	 * Read properties file from RobRio disk into a Properties object.
 	 * @return A Properties object.
@@ -65,42 +65,42 @@ public class Util
 	public static Properties readProperties() throws IOException
 	{
 		consoleLog();
-		
+
 		Properties props = new Properties();
-		
+
 		FileInputStream is = new FileInputStream("/home/lvuser/Robot.properties");
-		
+
 		props.load(is);
-		
+
 		is.close();
-		
+
 		//props.list(new PrintStream(System.out));
 		props.list(logPrintStream);
 
 		return props;
 	}
-	
+
 	/**
-	 * Configures and holds (static) classes for our custom logging system. 
+	 * Configures and holds (static) classes for our custom logging system.
 	 * Call setup() method to initialize logging. Logging is then done via
 	 * the logging convenience methods in the Util class.
 	 */
-	public static class CustomLogger 
+	public static class CustomLogger
 	{
         static private FileHandler 		fileTxt;
         //static private SimpleFormatter	formatterTxt;
         static private LogFormatter		logFormatter;
-        
+
         /**
          *  Initializes our logging system.
          *  Call before using any logging methods.
          *  @throws IOException
          */
-        static public void setup() throws IOException 
+        static public void setup() throws IOException
         {
             // get the global logger to configure it and add a file handler.
             Logger logger = Logger.getGlobal();
-                   
+
             logger.setLevel(Level.ALL);
 
             // If we decide to redirect system.out to our log handler, then following
@@ -108,12 +108,12 @@ public class Util
             // a recursive loop. We would only redirect system.out if we only want to
             // log to the file. If we delete the console handler we can skip setting
             // the formatter...otherwise we set our formatter on the console logger.
-            
+
             Logger rootLogger = Logger.getLogger("");
 
             Handler[] handlers = rootLogger.getHandlers();
-            
-//            if (handlers[0] instanceof ConsoleHandler) 
+
+//            if (handlers[0] instanceof ConsoleHandler)
 //            {
 //                rootLogger.removeHandler(handlers[0]);
 //                return;
@@ -125,49 +125,49 @@ public class Util
             if (handlers[0] instanceof ConsoleHandler) handlers[0].setFormatter(logFormatter);
 
             // Now create a handler to log to a file on roboRio "disk".
-            
+
             //if (true) throw new IOException("Test Exception");
-            
+
             if (new File("/home/lvuser/Logging.txt.99").exists() != true)
             	fileTxt = new FileHandler("/home/lvuser/Logging.txt", 0, 99);
             else
             	throw new IOException("Max number of log files reached.");
-            
+
             fileTxt.setFormatter(logFormatter);
 
             logger.addHandler(fileTxt);
         }
 	}
-	
+
 	// Our custom formatter for logging output.
-	
-	private static class LogFormatter extends Formatter 
+
+	private static class LogFormatter extends Formatter
 	{
         SimpleDateFormat dateFormat = new SimpleDateFormat("hh:mm:ss:S");
-        
+
         public LogFormatter()
         {
             dateFormat.setTimeZone(TimeZone.getTimeZone("America/Los_Angeles"));
         }
 
-        public String format(LogRecord rec) 
+        public String format(LogRecord rec)
         {
             StringBuffer buf = new StringBuffer(1024);
-            
+
             buf.append(String.format("<%d>", rec.getThreadID()));
             buf.append(dateFormat.format(new Date(rec.getMillis())));
             buf.append(" ");
             buf.append(formatMessage(rec));
             buf.append("\n");
-        
+
             return buf.toString();
         }
 	}
-	
+
 	// An output stream that writes to our logging system. Writes data with flush on
 	// flush call or on a newline character in the stream.
-	
-	private static class LoggingOutputStream extends OutputStream 
+
+	private static class LoggingOutputStream extends OutputStream
 	{
 	    private static final int	DEFAULT_BUFFER_LENGTH = 2048;
 	    private boolean 			hasBeenClosed = false;
@@ -181,22 +181,22 @@ public class Util
 	        count = 0;
 	    }
 
-	    public void write(final int b) throws IOException 
+	    public void write(final int b) throws IOException
 	    {
 	        if (hasBeenClosed) {throw new IOException("The stream has been closed.");}
-	        
+
 	        // don't log nulls
 	        if (b == 0) return;
-	        
+
 	        // force flush on newline character, dropping the newline.
-	        if ((byte) b == '\n') 
+	        if ((byte) b == '\n')
 	        {
 	        	flush();
 	        	return;
 	        }
-	        
+
 	        // would this be writing past the buffer?
-	        if (count == curBufLength) 
+	        if (count == curBufLength)
 	        {
 	            // grow the buffer
 	            final int newBufLength = curBufLength + DEFAULT_BUFFER_LENGTH;
@@ -207,27 +207,27 @@ public class Util
 	        }
 
 	        buf[count] = (byte) b;
-	        
+
 	        count++;
 	    }
 
 	    @SuppressWarnings("deprecation")
-		public void flush() 
+		public void flush()
 	    {
 	        if (count == 0) return;
-	        
+
 	        final byte[] bytes = new byte[count];
 
 	        System.arraycopy(buf, 0, bytes, 0, count);
-	        
+
 	        String str = new String(bytes);
-	        
+
 	        LCD.consoleLogNoFormat(str);
-	        
+
 	        count = 0;
 	    }
 
-	    public void close() 
+	    public void close()
 	    {
 	        flush();
 	        hasBeenClosed = true;
@@ -246,19 +246,19 @@ public class Util
     private static String currentMethod(Integer level)
     {
         StackTraceElement stackTrace[];
-    
+
         stackTrace = new Throwable().getStackTrace();
 
         // This scheme depends on having one level in the package name between
         // Team4450 and the class name, ie: Team4450.lib.Util.method. New levels
         // will require rewrite.
-        
+
         try
         {
             String method = stackTrace[level].toString().split("4450.")[1];
-            
+
             int startPos = method.indexOf(".") + 1;
-            
+
             return method.substring(startPos);
         }
         catch (Throwable e)
@@ -278,7 +278,7 @@ public class Util
 
 	// Works the same as LCD.consoleLog but automatically includes the program location from which
 	// trace was called.
-    
+
     /**
      * Write message to console log with optional formatting and program location.
      * @param message Message with optional format specifiers for listed parameters.
@@ -289,7 +289,7 @@ public class Util
 		// logs to the console as well as our log file on RR disk.
 		logger.log(Level.INFO, String.format("robot: %s: %s", currentMethod(2), String.format(message, parms)));
 	}
-    
+
 	/**
 	 * Write blank line with program location to the console log.
 	 */
@@ -298,18 +298,18 @@ public class Util
 		// logs to the console as well as our log file on RR disk.
 		logger.log(Level.INFO, String.format("robot: %s", currentMethod(2)));
 	}
-    
+
 	/**
 	 * Handler for uncaught exceptions. Records thread name and exception to log file.
 	 * @param t Thread where exception was thrown.
 	 * @param e Exception that was  thrown.
 	 */
-	public static void uncaughtException(Thread t, Throwable e) 
+	public static void uncaughtException(Thread t, Throwable e)
 	{
         consoleLog("Uncaught exception from thread " + t);
         logException(e);
     }
-    
+
 	/**
 	 * Write exception message to DS console window and exception stack trace to
 	 * log file.
@@ -318,17 +318,17 @@ public class Util
 	public static void logException(Throwable e)
 	{
 		DriverStation.reportError(e.toString(), false);
-		
+
 		e.printStackTrace(Util.logPrintStream);
 	}
 
 	/** helper routine to get last received message for a given ID */
-	private static long checkMessage(int fullId, int deviceID) 
+	private static long checkMessage(int fullId, int deviceID)
 	{
 		ByteBuffer targetID = ByteBuffer.allocateDirect(4);
 		ByteBuffer timeStamp = ByteBuffer.allocateDirect(4);
 
-		try 
+		try
 		{
 			targetID.clear();
 			targetID.order(ByteOrder.LITTLE_ENDIAN);
@@ -337,23 +337,23 @@ public class Util
 			timeStamp.clear();
 			timeStamp.order(ByteOrder.LITTLE_ENDIAN);
 			timeStamp.asIntBuffer().put(0,0x00000000);
-			
+
 			CANJNI.FRCNetCommCANSessionMuxReceiveMessage(targetID.asIntBuffer(), 0x1fffffff, timeStamp);
-		
+
 			long retval = timeStamp.getInt();
-			
-			retval &= 0xFFFFFFFF; /* undo sign-extension */ 
-			
+
+			retval &= 0xFFFFFFFF; /* undo sign-extension */
+
 			return retval;
 		} catch (Exception e) {return -1;}
 	}
-	
+
 	/** polls for received framing to determine if a device is present.
-	 *   This is meant to be used once initially (and not periodically) since 
+	 *   This is meant to be used once initially (and not periodically) since
 	 *   this steals cached messages from the robot API.
 	 * @return ArrayList of strings holding the names of devices we've found.
 	 */
-	public static ArrayList<String> listCANDevices() 
+	public static ArrayList<String> listCANDevices()
 	{
 		ArrayList<String> retval = new ArrayList<String>();
 
@@ -361,17 +361,17 @@ public class Util
 		long pdp0_timeStamp0; // only look for PDP at '0'
 		long []pcm_timeStamp0 = new long[63];
 		long []srx_timeStamp0 = new long[63];
-		
+
 		pdp0_timeStamp0 = checkMessage(0x08041400,0);
-		 
-		for(int i = 0; i < 63; ++i) 
+
+		for(int i = 0; i < 63; ++i)
 		{
 			pcm_timeStamp0[i] = checkMessage(0x09041400, i);
 			srx_timeStamp0[i] = checkMessage(0x02041400, i);
 		}
 
 		/* wait 200ms */
-		try 
+		try
 		{
 			Thread.sleep(200);
 		} catch (InterruptedException e) {Util.logException(e);}
@@ -380,11 +380,11 @@ public class Util
 		long pdp0_timeStamp1; // only look for PDP at '0'
 		long []pcm_timeStamp1 = new long[63];
 		long []srx_timeStamp1 = new long[63];
-		
+
 		pdp0_timeStamp1 = checkMessage(0x08041400,0);
-		
-		
-		for(int i = 0; i < 63; ++i) 
+
+
+		for(int i = 0; i < 63; ++i)
 		{
 			pcm_timeStamp1[i] = checkMessage(0x09041400, i);
 			srx_timeStamp1[i] = checkMessage(0x02041400, i);
@@ -394,18 +394,18 @@ public class Util
 		if( pdp0_timeStamp0 >= 0 && pdp0_timeStamp1 >= 0 && pdp0_timeStamp0 != pdp0_timeStamp1)
 			retval.add("PDP 0");
 
-		for(int i = 0; i < 63; ++i) 
+		for(int i = 0; i < 63; ++i)
 		{
 			if( pcm_timeStamp0[i] >= 0 && pcm_timeStamp1[i] >= 0 && pcm_timeStamp0[i] != pcm_timeStamp1[i])
 				retval.add("PCM " + i);
-		
+
 			if( srx_timeStamp0[i] >= 0 && srx_timeStamp1[i] >= 0 && srx_timeStamp0[i] != srx_timeStamp1[i])
 				retval.add("SRX " + i);
 		}
-		
+
 		return retval;
 	}
-	
+
 	/**
 	 * Check a double value to be within a min/max range.
 	 * @param value Value to test.
@@ -416,12 +416,12 @@ public class Util
 	public static boolean checkRange(double value, double min, double max)
 	{
 		if (min > max) throw new BoundaryException("min is greater than max");
-		
+
 		if (value < min || value > max) return false;
-		
+
 		return true;
 	}
-	
+
 	/**
 	 * Check a double value to be within a min/max range. Throws exception with error message if not.
 	 * @param value Value to test.
@@ -434,7 +434,7 @@ public class Util
 	{
 		if (!checkRange(value, min, max)) throw new IllegalArgumentException(errorMessage);
 	}
-	
+
 	/**
 	 * Check an integer value to be within a min/max range.
 	 * @param value Value to test.
@@ -445,12 +445,12 @@ public class Util
 	public static boolean checkRange(int value, int min, int max)
 	{
 		if (min > max) throw new BoundaryException("min is greater than max");
-		
+
 		if (value < min || value > max) return false;
-		
+
 		return true;
 	}
-	
+
 	/**
 	 * Check an integer value to be within a min/max range. Throws exception with error message if not.
 	 * @param value Value to test.
@@ -463,7 +463,7 @@ public class Util
 	{
 		if (!checkRange(value, min, max)) throw new IllegalArgumentException(errorMessage);
 	}
-	
+
 	/**
 	 * Check an integer value to be in the range -magnitude to +magnitude.
 	 * @param value Value to test.
@@ -474,7 +474,7 @@ public class Util
 	{
 		return checkRange(value, -magnitude, +magnitude);
 	}
-	
+
 	/**
 	 * Check an integer value to be in the range -magnitude to +magnitude. Throws exception with error message if not.
 	 * @param value Value to test.
@@ -486,7 +486,7 @@ public class Util
 	{
 		if (!checkRange(value,magnitude)) throw new IllegalArgumentException(errorMessage);
 	}
-	
+
 	/**
 	 * Check a double value to be in the range -magnitude to +magnitude.
 	 * @param value Value to test.
@@ -497,7 +497,7 @@ public class Util
 	{
 		return checkRange(value, -magnitude, +magnitude);
 	}
-	
+
 	/**
 	 * Check a double value to be in the range -magnitude to +magnitude. Throws exception with error message if not.
 	 * @param value Value to test.
@@ -521,7 +521,7 @@ public class Util
 	{
 		return Math.max(Math.min(value, max), min);
 	}
-	
+
 	/**
 	 * Constrain an integer value to be in the range -magnitude to +magnitude.
 	 * @param value Value to test
@@ -544,7 +544,7 @@ public class Util
 	{
 		return Math.min(Math.max(value, min), max);
 	}
-	
+
 	/**
 	 * Constrain an double value to be in the range -magnitude to +magnitude.
 	 * @param value Value to test
@@ -555,7 +555,7 @@ public class Util
 	{
 		return clampValue(value, -magnitude, magnitude);
 	}
-	
+
 	/**
 	 * Round a double to a specified number of decimal places.
 	 * @param number Value to round.
@@ -563,13 +563,13 @@ public class Util
 	 * @param rounding Selected RoundingMode.
 	 * @return Rounded value.
 	 */
-	public static double round(double number, int decimalPlaces, RoundingMode rounding) 
+	public static double round(double number, int decimalPlaces, RoundingMode rounding)
 	{
 		BigDecimal bd = new BigDecimal(number);
 		bd = bd.setScale(decimalPlaces, rounding);
-		return bd.floatValue();
+		return bd.doubleValue();
 	}
-	
+
 	/**
 	 * Return the elapsed time since the last call to this method.
 	 * @return Elapsed time in seconds.
@@ -582,7 +582,7 @@ public class Util
 		timeMarker = now;
 		return elapsedTime;
 	}
-	
+
 	/**
 	 * Return the elapsed time between the specified timestamp and current timestamp.
 	 * @param previousTime Previous timestamp.
@@ -601,7 +601,7 @@ public class Util
 	{
 		return Timer.getFPGATimestamp();
 	}
-	
+
 	/**
 	 * Convert inches to meters.
 	 * @param inches Inches value to convert.
